requirements.txt
gpiozero
python-dotenv
requests